Ensure agent ids array is type int

+++ sagesim/agent.py
@@ -272,7 +272,7 @@ def contextualize_agent_data_tensors(
             agents_index_in_subcontext.append(int(index_in_subcontext))
             index_in_subcontext += 1
     return (
-        np.array(agent_ids_in_subcontext),
+        np.array(agent_ids_in_subcontext).astype(int),
         np.array(agents_index_in_subcontext),
         agent_data_tensors_subcontext,
     )
